requirements.txt
websockets
requests
yfinance
PyYAML
pandas
numpy
discord.py

